WIP. Added best_orders_tests. Preparing for best_orders_v2 RPC.
